First steps into a functionally complete app.
Emulates the first few steps of the program (still doesn’t verify
manifest). At the current point, it is a little mixed up from the
transition from python 2 to 3, but otherwise the only thing to look out
for is the referencing of local files. Use config.directory. Feed main,
config, mailController and conversionController into Platypus.